Resolve SonarCloud findings: split oversized modules, reduce complexity
